Fixed: Running cloned template results in error
When template linear_regression_manual is run it ends up with an error.
IMO This is not convenient especially when advicing to add print to
learn about dataset.

Error is caused by calling format on None. Fixed by adding a check.

To be honest, mostly I want to learn how to contribute to break the
entry barrier and ease up contributing later on.