Change syntax for getting 0th column
In my solution I had to modify this part because for some reason it didn't work for me... It seemed that the problem was slicing (e.g. train_data[:,0]), maybe it doesn't work with my python version (my python version is 3.10.7). Either way, I thought that if because of this change the code can be run on newer versions and also on the older versions then maybe it would be better this way. It is still a small piece of code thanks to python comprehensions. I know it's only a small change, but it might save someone a few seconds.