Rename the arguments parsed in __main__ from args to main_args.
The variable in `__main__` is in fact a global variable. When
it was just `args`, it could be confused with `args` given
to `main`, and sometimes people forgot to pass `args` to called
methods, because `args` was magically available everywhere.
However, that is not true in ReCodEx, where `__main__` is not executed.